old-projects/carga-dados-bigtable-manual-main/requirements.txt
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
google-api-core==2.11.0
google-auth==2.18.1
google-cloud-bigtable==2.18.1
google-cloud-core==2.3.2
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
proto-plus==1.22.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.30.0
rsa==4.9
six==1.16.0
urllib3==1.26.15
